debug uninstall

+++ inc/class-undisclosedinstall.php
@@ -10,16 +10,17 @@
 // ----------------------------------------
 
 if ( ! class_exists('UndisclosedInstall') ) :
+
 class UndisclosedInstall {
 	
 	// --------------------------------------------------
 	// de-/activation/uninstall hooks
 	// --------------------------------------------------
 	static function activate( ) {
+		global $wpdb;
+
 		if ( ! current_user_can( 'activate_plugins' ) )
 			return;
-			
-		global $wpdb;
 
 		self::_install_capabilities_table( );
 		
@@ -32,20 +33,17 @@ class UndisclosedInstall {
 		} else {
 			self::_install_posts_table( );
 		}
-		
 	}
 	static function deactivate( $networkwide ) {
 		if ( ! current_user_can( 'activate_plugins' ) )
 			return;
-
-		
+//		self::uninstall();
 	}
 	static function uninstall() {
 		if ( ! current_user_can( 'activate_plugins' ) )
 			return;
 		
-		self::_remove_custom_caps();
-		
+		self::_uninstall_custom_caps();
 		self::_uninstall_capabilities_table();
 		
 		if (function_exists('is_multisite') && is_multisite() ) {
@@ -54,10 +52,10 @@ class UndisclosedInstall {
 				switch_to_blog($blog_id);
 				self::_uninstall_posts_table( );
 			}
+			restore_current_blog();
 		} else {
 			self::_uninstall_posts_table( );
 		}
-		restore_currnet_blog();
 	}
 	
 
@@ -66,21 +64,29 @@ class UndisclosedInstall {
 	// --------------------------------------------------
 	private static function _install_posts_table( ) {
 		global $wpdb;
-		$wpdb->query("ALTER IGNORE TABLE $wpdb->posts ADD COLUMN post_comment_cap varchar(128) NOT NULL DEFAULT 'exist' AFTER `post_status`;");
-		$wpdb->query("ALTER IGNORE TABLE $wpdb->posts ADD COLUMN post_edit_cap varchar(128) NOT NULL DEFAULT 'exist' AFTER `post_status`;");
-		$wpdb->query("ALTER IGNORE TABLE $wpdb->posts ADD COLUMN post_view_cap varchar(128) NOT NULL DEFAULT 'exist' AFTER `post_status`;");
-		$wpdb->query("ALTER IGNORE TABLE $wpdb->posts ADD INDEX `view_cap` (`post_view_cap`);");
-		$wpdb->query("ALTER IGNORE TABLE $wpdb->posts ADD INDEX `edit_cap` (`post_edit_cap`);");
-		$wpdb->query("ALTER IGNORE TABLE $wpdb->posts ADD INDEX `comment_cap` (`post_comment_cap`);");
+		$cols = array( 'comment_cap'=>'post_comment_cap' , 'edit_cap'=>'post_edit_cap' , 'view_cap'=>'post_view_cap' );
+		foreach ( $cols as $idx => $col ) {
+			$c = $wpdb->get_results("SHOW COLUMNS FROM $wpdb->posts LIKE '$col'");
+			if (empty($c))
+				$wpdb->query("ALTER TABLE $wpdb->posts ADD COLUMN $col varchar(128) NOT NULL DEFAULT 'exist' AFTER `post_status`;");
+				
+			$i = $wpdb->query("SHOW INDEX FROM $wpdb->posts WHERE Key_name = '$idx'");
+			if (empty($i))
+				$wpdb->query("ALTER TABLE $wpdb->posts ADD INDEX `$idx` (`$col`);");
+		}
 	}
 	private static function _uninstall_posts_table( ) {
 		global $wpdb;
-		$wpdb->query("ALTER IGNORE TABLE $wpdb->posts DROP COLUMN post_comment_cap;");
-		$wpdb->query("ALTER IGNORE TABLE $wpdb->posts DROP COLUMN post_edit_cap;");
-		$wpdb->query("ALTER IGNORE TABLE $wpdb->posts DROP COLUMN post_view_cap;");
-		$wpdb->query("ALTER IGNORE TABLE $wpdb->posts DROP INDEX (`view_cap`);");
-		$wpdb->query("ALTER IGNORE TABLE $wpdb->posts DROP INDEX (`edit_cap`);");
-		$wpdb->query("ALTER IGNORE TABLE $wpdb->posts DROP INDEX (`comment_cap`);");
+		$cols = array( 'comment_cap'=>'post_comment_cap' , 'edit_cap'=>'post_edit_cap' , 'view_cap'=>'post_view_cap' );
+		foreach ( $cols as $idx => $col ) {
+			$c = $wpdb->get_results("SHOW COLUMNS FROM $wpdb->posts LIKE '$col'");
+			if (!empty($c))
+				$wpdb->query("ALTER TABLE $wpdb->posts DROP COLUMN $col;");
+				
+			$i = $wpdb->query("SHOW INDEX FROM $wpdb->posts WHERE Key_name = '$idx'");
+			if (!empty($i))
+				$wpdb->query("ALTER TABLE $wpdb->posts DROP INDEX ('$idx');");
+		}
 	}
 	
 	
@@ -90,7 +96,7 @@ class UndisclosedInstall {
 	private static function _install_capabilities_table( ) {
 		global $wpdb;
 		$table_name = $wpdb->base_prefix . WPUND_USERLABEL_TABLE;
-		if($wpdb->get_var("show tables like '$table_name'") != $table_name) {
+		if ($wpdb->get_var("show tables like '$table_name'") != $table_name) {
 			$sql = "CREATE TABLE " . $table_name . " (
 				ID bigint(20) unsigned NOT NULL AUTO_INCREMENT,
 				cap_title varchar(64) NOT NULL,
@@ -110,22 +116,29 @@ class UndisclosedInstall {
 	private static function _uninstall_capabilities_table( ) {
 		global $wpdb;
 		$table_name = $wpdb->base_prefix . WPUND_USERLABEL_TABLE;
-		$wpdb->query("DROP TABLE IF EXISTS $table_name");
+		if ($wpdb->get_var("show tables like '$table_name'") == $table_name)
+			$wpdb->query("DROP TABLE IF EXISTS $table_name");
 	}
 	// --------------------------------------------
 	// remove Caps from User
 	private function _uninstall_custom_caps( ) {
 		global $wpdb;
 		$table_name = $wpdb->base_prefix . WPUND_USERLABEL_TABLE;
-		$uids = $wpdb->get_col( "SELECT user_id FROM $wpdb->usermeta WHERE meta_key LIKE '{$wpdb->base_prefix}%capabilities' AND meta_value LIKE '%\"".WPUND_USERLABEL_PREFIX."%'" );
-		$caps = $wpdb->get_col( "SELECT capability FROM $table_name" );
-		foreach ( $uids as $uid) {
-			$user = new WP_User( $uid );
-			foreach ( $caps as $capability )
-				if ( $user->has_cap() )
-					$user->remove_cap( $capability );
+
+		$query =  "SELECT * FROM $wpdb->usermeta WHERE meta_key LIKE '{$wpdb->base_prefix}%capabilities' AND meta_value LIKE '%\"".WPUND_USERLABEL_PREFIX."%'" ;
+		$usermeta = $wpdb->get_results($query);
+		foreach ( $usermeta as $meta) {
+			$caps = maybe_unserialize($meta->meta_value);
+			foreach ( array_keys($caps) as $key ) 
+				if ( strpos( $key , WPUND_USERLABEL_PREFIX ) === 0 )
+					unset($caps[$key]);
+			$wpdb->query( $wpdb->prepare( "UPDATE $wpdb->usermeta SET meta_value=%s WHERE umeta_id=%d",
+				serialize( $caps ), 
+				$meta->umeta_id
+			) );
 		}
-		
+		$wpdb->query( "DELETE FROM $wpdb->usermeta WHERE meta_key = '".WPUND_GLOBAL_USERMETA_KEY."'" );
+
 	}
 	
 }
